+++ cmd/stakercli/daemon/daemoncommands.go
@@ -4,13 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"strconv"
 
 	"github.com/babylonlabs-io/btc-staker/cmd/stakercli/helpers"
-	scfg "github.com/babylonlabs-io/btc-staker/stakercfg"
 	dc "github.com/babylonlabs-io/btc-staker/stakerservice/client"
-	"github.com/babylonlabs-io/networks/parameters/parser"
-	"github.com/cometbft/cometbft/libs/os"
 	"github.com/urfave/cli"
 )
 
@@ -36,17 +32,11 @@ var DaemonCommands = []cli.Command{
 }
 
 const (
-	stakingDaemonAddressFlag   = "daemon-address"
 	offsetFlag                 = "offset"
 	limitFlag                  = "limit"
 	fpPksFlag                  = "finality-providers-pks"
 	stakingTransactionHashFlag = "staking-transaction-hash"
 	stakerAddressFlag          = "staker-address"
-	txInclusionHeightFlag      = "tx-inclusion-height"
-)
-
-var (
-	defaultStakingDaemonAddress = "tcp://127.0.0.1:" + strconv.Itoa(scfg.DefaultRPCPort)
 )
 
 var checkDaemonHealthCmd = cli.Command{
@@ -55,9 +45,9 @@ var checkDaemonHealthCmd = cli.Command{
 	Usage:     "Check if staker daemon is running.",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "Full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 	},
 	Action: checkHealth,
@@ -69,9 +59,9 @@ var listOutputsCmd = cli.Command{
 	Usage:     "List unspent outputs in connected wallet.",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "Full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 	},
 	Action: listOutputs,
@@ -83,9 +73,9 @@ var babylonFinalityProvidersCmd = cli.Command{
 	Usage:     "List current BTC finality providers on Babylon chain",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.IntFlag{
 			Name:  offsetFlag,
@@ -107,9 +97,9 @@ var stakeCmd = cli.Command{
 	Usage:     "Stake an amount of BTC to Babylon",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.StringFlag{
 			Name:     stakerAddressFlag,
@@ -142,14 +132,14 @@ var stakeCmd = cli.Command{
 var stakeFromPhase1Cmd = cli.Command{
 	Name:      "stake-from-phase1",
 	ShortName: "stfp1",
-	Usage: "\nstakercli daemon stake-from-phase1 [fullpath/to/global_parameters.json]" +
+	Usage: "\nstakercli daemon stake-from-phase1" +
 		" --staking-transaction-hash [txHashHex] --staker-address [btcStakerAddrHex] --tx-inclusion-height [blockHeightTxInclusion]",
 	Description: "Creates a Babylon BTC delegation transaction from the Phase1 BTC staking transaction",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.StringFlag{
 			Name:     stakingTransactionHashFlag,
@@ -162,7 +152,7 @@ var stakeFromPhase1Cmd = cli.Command{
 			Required: true,
 		},
 		cli.Uint64Flag{
-			Name:  txInclusionHeightFlag,
+			Name:  helpers.TxInclusionHeightFlag,
 			Usage: "Expected BTC height at which transaction was included. This value is important to choose correct global parameters for transaction, if set doesn't query bitcoin to get the block height from txHash",
 		},
 	},
@@ -175,9 +165,9 @@ var unstakeCmd = cli.Command{
 	Usage:     "Spends staking transaction and sends funds back to staker; this can only be done after timelock of staking transaction expires",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.StringFlag{
 			Name:     stakingTransactionHashFlag,
@@ -194,9 +184,9 @@ var unbondCmd = cli.Command{
 	Usage:     "initiates unbonding flow: build unbonding tx, send to babylon, wait for signatures, and send unbonding tx to bitcoin",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.StringFlag{
 			Name:     stakingTransactionHashFlag,
@@ -213,9 +203,9 @@ var stakingDetailsCmd = cli.Command{
 	Usage:     "Displays details of staking transaction with given hash",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.StringFlag{
 			Name:     stakingTransactionHashFlag,
@@ -232,9 +222,9 @@ var listStakingTransactionsCmd = cli.Command{
 	Usage:     "List current staking transactions in db",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.IntFlag{
 			Name:  offsetFlag,
@@ -256,9 +246,9 @@ var withdrawableTransactionsCmd = cli.Command{
 	Usage:     "List current tranactions that can be withdrawn i.e funds can be transferred back to staker address",
 	Flags: []cli.Flag{
 		cli.StringFlag{
-			Name:  stakingDaemonAddressFlag,
+			Name:  helpers.StakingDaemonAddressFlag,
 			Usage: "full address of the staker daemon in format tcp:://<host>:<port>",
-			Value: defaultStakingDaemonAddress,
+			Value: helpers.DefaultStakingDaemonAddress,
 		},
 		cli.IntFlag{
 			Name:  offsetFlag,
@@ -275,7 +265,7 @@ var withdrawableTransactionsCmd = cli.Command{
 }
 
 func checkHealth(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -295,7 +285,7 @@ func checkHealth(ctx *cli.Context) error {
 }
 
 func listOutputs(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -315,7 +305,7 @@ func listOutputs(ctx *cli.Context) error {
 }
 
 func babylonFinalityProviders(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -347,7 +337,7 @@ func babylonFinalityProviders(ctx *cli.Context) error {
 }
 
 func stake(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -372,7 +362,7 @@ func stake(ctx *cli.Context) error {
 }
 
 func stakeFromPhase1TxBTC(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -384,21 +374,7 @@ func stakeFromPhase1TxBTC(ctx *cli.Context) error {
 		return errors.New("staking tx hash hex is empty")
 	}
 
-	inputGlobalParamsFilePath := ctx.Args().First()
-	if len(inputGlobalParamsFilePath) == 0 {
-		return errors.New("json file input is empty")
-	}
-
-	if !os.FileExists(inputGlobalParamsFilePath) {
-		return fmt.Errorf("json file input %s does not exist", inputGlobalParamsFilePath)
-	}
-
-	globalParams, err := parser.NewParsedGlobalParamsFromFile(inputGlobalParamsFilePath)
-	if err != nil {
-		return fmt.Errorf("error parsing file %s: %w", inputGlobalParamsFilePath, err)
-	}
-
-	blockHeighTxInclusion := ctx.Uint64(txInclusionHeightFlag)
+	blockHeighTxInclusion := ctx.Uint64(helpers.TxInclusionHeightFlag)
 	if blockHeighTxInclusion == 0 {
 		resp, err := client.BtcTxDetails(sctx, stakingTransactionHash)
 		if err != nil {
@@ -408,18 +384,22 @@ func stakeFromPhase1TxBTC(ctx *cli.Context) error {
 		blockHeighTxInclusion = uint64(resp.Blk.Height)
 	}
 
-	paramsForHeight := globalParams.GetVersionedGlobalParamsByHeight(blockHeighTxInclusion)
-	if paramsForHeight == nil {
-		return fmt.Errorf("error getting param version from global params %s with height %d", inputGlobalParamsFilePath, blockHeighTxInclusion)
+	respParamsByHeight, err := client.BtcStakingParamByBtcHeight(sctx, uint32(blockHeighTxInclusion))
+	if err != nil {
+		return fmt.Errorf("failed to get btc staking parameters: %w", err)
 	}
+	btcStakingParams := respParamsByHeight.StakingParams
 
 	stakerAddress := ctx.String(stakerAddressFlag)
-	_, err = client.BtcDelegationFromBtcStakingTx(sctx, stakerAddress, stakingTransactionHash, paramsForHeight)
-	return err
+	_, err = client.BtcDelegationFromBtcStakingTx(sctx, stakerAddress, stakingTransactionHash, btcStakingParams.CovenantPks, btcStakingParams.CovenantQuorum)
+	if err != nil {
+		return fmt.Errorf("failed to delegate from btc staking tx: %w", err)
+	}
+	return nil
 }
 
 func unstake(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -440,7 +420,7 @@ func unstake(ctx *cli.Context) error {
 }
 
 func unbond(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -461,7 +441,7 @@ func unbond(ctx *cli.Context) error {
 }
 
 func stakingDetails(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -482,7 +462,7 @@ func stakingDetails(ctx *cli.Context) error {
 }
 
 func listStakingTransactions(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err
@@ -514,7 +494,7 @@ func listStakingTransactions(ctx *cli.Context) error {
 }
 
 func withdrawableTransactions(ctx *cli.Context) error {
-	daemonAddress := ctx.String(stakingDaemonAddressFlag)
+	daemonAddress := ctx.String(helpers.StakingDaemonAddressFlag)
 	client, err := dc.NewStakerServiceJSONRPCClient(daemonAddress)
 	if err != nil {
 		return err